refactor(parsing): split merchant providers
