+++ src/parsing/merchants/identifier-regex.ts
@@ -1,0 +1,4 @@
+export const getIdentifierFromRegex =
+  (regex: RegExp) =>
+  (input: string): string | null =>
+    input.match(regex)?.groups?.identifier ?? null
