+++ src/parsing/merchants/index.ts
@@ -1,9 +1,8 @@
-import {
-  boltzSwapDomains,
-  getBoltzSwapIdentifier,
-  getBoltzSwapMerchants,
-} from "./boltz-swap-merchants"
+import { boltzSwapMerchant } from "./boltz-swap-merchants"
+import { moneyBadgerMerchants } from "./money-badger-merchants"
 import type { Network } from "../types"
+
+export { getIdentifierFromRegex } from "./identifier-regex"
 
 export type MerchantCategory = "merchant-payment" | "swap"
 
@@ -18,230 +17,14 @@ export type Merchant = {
   displayCurrency?: string
 }
 
-type MerchantConfig = Omit<Merchant, "lnurl"> & {
+export type MerchantConfig = Omit<Merchant, "lnurl"> & {
   getIdentifier: (input: string) => string | null
   defaultDomain: string
   domains: { [K in Network]: string }
   getMerchants?: (identifier: string, network: Network) => Merchant[]
 }
 
-export const getIdentifierFromRegex =
-  (regex: RegExp) =>
-  (input: string): string | null =>
-    input.match(regex)?.groups?.identifier ?? null
-
-const moneyBadgerTermsUrl = "https://www.moneybadger.co.za/deals/terms-and-conditions"
-const moneyBadgerMerchant = {
-  category: "merchant-payment",
-  description: "Money Badger merchant",
-  companyName: "Money Badger",
-  termsUrl: moneyBadgerTermsUrl,
-} satisfies Pick<Merchant, "category" | "description" | "companyName" | "termsUrl">
-
-export const merchants: MerchantConfig[] = [
-  {
-    id: "picknpay",
-    title: "Pick n Pay",
-    ...moneyBadgerMerchant,
-    displayCurrency: "ZAR",
-    getIdentifier: getIdentifierFromRegex(
-      /(?<identifier>.*za\.co\.electrum\.picknpay.*)/iu,
-    ),
-    defaultDomain: "cryptoqr.net",
-    domains: {
-      mainnet: "cryptoqr.net",
-      signet: "staging.cryptoqr.net",
-      regtest: "staging.cryptoqr.net",
-    },
-  },
-  {
-    id: "ecentric",
-    title: "Ecentric",
-    ...moneyBadgerMerchant,
-    displayCurrency: "ZAR",
-    getIdentifier: getIdentifierFromRegex(/(?<identifier>.*za\.co\.ecentric.*)/iu),
-    defaultDomain: "cryptoqr.net",
-    domains: {
-      mainnet: "cryptoqr.net",
-      signet: "staging.cryptoqr.net",
-      regtest: "staging.cryptoqr.net",
-    },
-  },
-  {
-    id: "yoyo",
-    title: "Yoyo",
-    ...moneyBadgerMerchant,
-    displayCurrency: "ZAR",
-    getIdentifier: getIdentifierFromRegex(
-      /(?<identifier>.*(wigroup\.co|yoyogroup\.co).*)/iu,
-    ),
-    defaultDomain: "cryptoqr.net",
-    domains: {
-      mainnet: "cryptoqr.net",
-      signet: "staging.cryptoqr.net",
-      regtest: "staging.cryptoqr.net",
-    },
-  },
-  {
-    id: "zapper",
-    title: "Zapper",
-    ...moneyBadgerMerchant,
-    displayCurrency: "ZAR",
-    getIdentifier: getIdentifierFromRegex(
-      /(?<identifier>.*(zapper\.com|\d+\.zap\.pe).*)/iu,
-    ),
-    defaultDomain: "cryptoqr.net",
-    domains: {
-      mainnet: "cryptoqr.net",
-      signet: "staging.cryptoqr.net",
-      regtest: "staging.cryptoqr.net",
-    },
-  },
-  {
-    id: "payat",
-    title: "Payat",
-    ...moneyBadgerMerchant,
-    displayCurrency: "ZAR",
-    getIdentifier: getIdentifierFromRegex(/(?<identifier>.*payat\.io.*)/iu),
-    defaultDomain: "cryptoqr.net",
-    domains: {
-      mainnet: "cryptoqr.net",
-      signet: "staging.cryptoqr.net",
-      regtest: "staging.cryptoqr.net",
-    },
-  },
-  {
-    id: "paynow-netcash",
-    title: "Paynow Netcash",
-    ...moneyBadgerMerchant,
-    displayCurrency: "ZAR",
-    getIdentifier: getIdentifierFromRegex(/(?<identifier>.*paynow\.netcash\.co\.za.*)/iu),
-    defaultDomain: "cryptoqr.net",
-    domains: {
-      mainnet: "cryptoqr.net",
-      signet: "staging.cryptoqr.net",
-      regtest: "staging.cryptoqr.net",
-    },
-  },
-  {
-    id: "paynow-sagepay",
-    title: "Paynow Sagepay",
-    ...moneyBadgerMerchant,
-    displayCurrency: "ZAR",
-    getIdentifier: getIdentifierFromRegex(/(?<identifier>.*paynow\.sagepay\.co\.za.*)/iu),
-    defaultDomain: "cryptoqr.net",
-    domains: {
-      mainnet: "cryptoqr.net",
-      signet: "staging.cryptoqr.net",
-      regtest: "staging.cryptoqr.net",
-    },
-  },
-  {
-    id: "standard-bank-scantopay",
-    title: "Standard Bank Scantopay",
-    ...moneyBadgerMerchant,
-    displayCurrency: "ZAR",
-    getIdentifier: getIdentifierFromRegex(/(?<identifier>SK-\d{1,}-\d{23})/iu),
-    defaultDomain: "cryptoqr.net",
-    domains: {
-      mainnet: "cryptoqr.net",
-      signet: "staging.cryptoqr.net",
-      regtest: "staging.cryptoqr.net",
-    },
-  },
-  {
-    id: "transactionjunction",
-    title: "Transaction Junction",
-    ...moneyBadgerMerchant,
-    displayCurrency: "ZAR",
-    getIdentifier: getIdentifierFromRegex(
-      /(?<identifier>.*transactionjunction\.co\.za.*)/iu,
-    ),
-    defaultDomain: "cryptoqr.net",
-    domains: {
-      mainnet: "cryptoqr.net",
-      signet: "staging.cryptoqr.net",
-      regtest: "staging.cryptoqr.net",
-    },
-  },
-  {
-    id: "servest-parking",
-    title: "Servest Parking",
-    ...moneyBadgerMerchant,
-    displayCurrency: "ZAR",
-    getIdentifier: getIdentifierFromRegex(/(?<identifier>CRSTPC-\d+-\d+-\d+-\d+-\d+)/iu),
-    defaultDomain: "cryptoqr.net",
-    domains: {
-      mainnet: "cryptoqr.net",
-      signet: "staging.cryptoqr.net",
-      regtest: "staging.cryptoqr.net",
-    },
-  },
-  {
-    id: "payat-generic",
-    title: "Payat",
-    ...moneyBadgerMerchant,
-    displayCurrency: "ZAR",
-    getIdentifier: getIdentifierFromRegex(/(?<identifier>.{2}\/.{4}\/.{20})/iu),
-    defaultDomain: "cryptoqr.net",
-    domains: {
-      mainnet: "cryptoqr.net",
-      signet: "staging.cryptoqr.net",
-      regtest: "staging.cryptoqr.net",
-    },
-  },
-  {
-    id: "scantopay-url",
-    title: "Scantopay Url",
-    ...moneyBadgerMerchant,
-    displayCurrency: "ZAR",
-    getIdentifier: getIdentifierFromRegex(/(?<identifier>.*(scantopay\.io).*)/iu),
-    defaultDomain: "cryptoqr.net",
-    domains: {
-      mainnet: "cryptoqr.net",
-      signet: "staging.cryptoqr.net",
-      regtest: "staging.cryptoqr.net",
-    },
-  },
-  {
-    id: "scantopay-10-digits",
-    title: "Scantopay 10 Digits",
-    ...moneyBadgerMerchant,
-    displayCurrency: "ZAR",
-    getIdentifier: getIdentifierFromRegex(/^(?<identifier>\d{10})$/iu),
-    defaultDomain: "cryptoqr.net",
-    domains: {
-      mainnet: "cryptoqr.net",
-      signet: "staging.cryptoqr.net",
-      regtest: "staging.cryptoqr.net",
-    },
-  },
-  {
-    id: "snapscan",
-    title: "Snapscan",
-    ...moneyBadgerMerchant,
-    displayCurrency: "ZAR",
-    getIdentifier: getIdentifierFromRegex(/(?<identifier>.*(snapscan).*)/iu),
-    defaultDomain: "cryptoqr.net",
-    domains: {
-      mainnet: "cryptoqr.net",
-      signet: "staging.cryptoqr.net",
-      regtest: "staging.cryptoqr.net",
-    },
-  },
-  {
-    id: "blink-boltz-swap",
-    category: "swap",
-    title: "Boltz Swap",
-    description: "Swap sats with Boltz",
-    companyName: "Boltz",
-    termsUrl: "https://boltz.exchange/terms",
-    getIdentifier: getBoltzSwapIdentifier,
-    defaultDomain: boltzSwapDomains.mainnet,
-    domains: boltzSwapDomains,
-    getMerchants: getBoltzSwapMerchants,
-  },
-]
+export const merchants: MerchantConfig[] = [...moneyBadgerMerchants, boltzSwapMerchant]
 
 // strict URI encode adhering to RFC 3986
 export const strictUriEncode = (uriComponent: string | number | boolean): string => {
